+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Typo in top padding for title
+
 ## [v0.6.2]
 ### Fixed
 - Encoding issue in report.
@@ -34,8 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v0.5.5]
 ### Changed
 - Bootstrap alerts now a section that can be added to a report.
-### Added 
-- disable SI prefixes on percentage values in graphics.infographic 
+### Added
+- disable SI prefixes on percentage values in graphics.infographic
 
 ## [v0.5.4]
 ### Changed