+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 ### Fixed
 - Typo in top padding for title
+- Nextclade web component version bump to 1.0.2
 
 ## [v0.6.2]
 ### Fixed