+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [v0.6.11]
 ### Changed
 - title lead can be customised
-- enable html escaping to be turned off in the table function 
+- enable html escaping to be turned off in the table function
 
 ## [v0.6.10]
 ### Fixed